requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.115.7
joblib==1.4.2
openai==1.60.0
pydantic==2.10.6
python-dotenv==1.0.1
supabase==2.11.0
superpowered-sdk==0.1.43
termcolor==2.5.0
uvicorn==0.34.0
